requirements.txt
tensorflow==2.20.0
matplotlib==3.10.3
pyyaml==6.0.2
dvc[gs]==3.60.1
numpy>=1.26.0
scikit-learn>=1.0.0
seaborn>=0.11.0
mlflow>=2.8.0
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.38.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.0.0
httpx>=0.24.0